backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# LangChain & LangGraph - Let pip resolve compatible versions
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langsmith>=0.2.0,<0.3.0

# Resolve tiktoken conflict
tiktoken>=0.7,<1

# Vector Store
pinecone-client==3.0.0

# APIs & Tools
tavily-python>=0.3.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6

# Google APIs
google-auth>=2.48.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.3.0
google-api-python-client>=2.188.0

# WebSocket
websockets==12.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
